requirements.txt
streamlit
gradio
langchain
langchain_ollama
PyMuPDF
sentence-transformers
python-dotenv
PyMuPDF
hdbcli
numpy
tiktoken